--- a/hbase-balancer/src/test/java/org/apache/hadoop/hbase/master/balancer/TestStochasticLoadBalancer.java
+++ b/hbase-balancer/src/test/java/org/apache/hadoop/hbase/master/balancer/TestStochasticLoadBalancer.java
@@ -235,6 +235,7 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
     BalancerClusterState clusterState = mockCluster(cluster);
     Map<TableName, Map<ServerName, List<RegionInfo>>> LoadOfAllTable =
       (Map) mockClusterServersWithTables(servers);
+    boolean oldIsByTable = conf.getBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, false);
     try {
       boolean[] perTableBalancerConfigs = { true, false };
       for (boolean isByTable : perTableBalancerConfigs) {
@@ -259,7 +260,7 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
         assertEquals(curOverallCost, curOverallCostInMetrics, 0.001);
       }
     } finally {
-      conf.unset(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE);
+      conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, oldIsByTable);
       loadBalancer.onConfigurationChange(conf);
     }
   }
@@ -267,6 +268,7 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
   @Test
   public void testUpdateStochasticCosts() {
     float minCost = conf.getFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 0.05f);
+    boolean oldIsByTable = conf.getBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, false);
     try {
       int[] cluster = new int[] { 10, 0 };
       Map<ServerName, List<RegionInfo>> servers = mockClusterServers(cluster);
@@ -289,7 +291,7 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
     } finally {
       // reset config
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", minCost);
-      conf.unset(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE);
+      conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, oldIsByTable);
       loadBalancer.onConfigurationChange(conf);
     }
   }
@@ -297,6 +299,7 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
   @Test
   public void testUpdateStochasticCostsIfBalanceNotRan() {
     float minCost = conf.getFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 0.05f);
+    boolean oldIsByTable = conf.getBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, false);
     try {
       int[] cluster = new int[] { 10, 10 };
       Map<ServerName, List<RegionInfo>> servers = mockClusterServers(cluster);
@@ -319,17 +322,22 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
     } finally {
       // reset config
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", minCost);
-      conf.unset(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE);
+      conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, oldIsByTable);
       loadBalancer.onConfigurationChange(conf);
     }
   }
 
   @Test
   public void testNeedBalance() {
+    boolean oldIsByTable = conf.getBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, false);
     float minCost = conf.getFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 0.05f);
     float slop = conf.getFloat(HConstants.LOAD_BALANCER_SLOP_KEY, 0.2f);
+    boolean runMaxSteps = conf.getBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+    long maxSteps = conf.getLong("hbase.master.balancer.stochastic.maxSteps", 1000000);
     conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 1.0f);
     conf.setFloat(HConstants.LOAD_BALANCER_SLOP_KEY, -1f);
+    conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+    conf.setLong("hbase.master.balancer.stochastic.maxSteps", 5000L);
     try {
       // Test with/without per table balancer.
       boolean[] perTableBalancerConfigs = { true, false };
@@ -342,9 +350,11 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
       }
     } finally {
       // reset config
-      conf.unset(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE);
+      conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, oldIsByTable);
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", minCost);
       conf.setFloat(HConstants.LOAD_BALANCER_SLOP_KEY, slop);
+      conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", runMaxSteps);
+      conf.setLong("hbase.master.balancer.stochastic.maxSteps", maxSteps);
       loadBalancer.onConfigurationChange(conf);
     }
   }
@@ -352,6 +362,8 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
   @Test
   public void testBalanceOfSloppyServers() {
     float minCost = conf.getFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 0.025f);
+    boolean runMaxSteps = conf.getBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+    long maxSteps = conf.getLong("hbase.master.balancer.stochastic.maxSteps", 1000000);
     try {
       // We are testing slop checks, so don't "accidentally" balance due to a minCost calculation.
       // During development, imbalance of a 100 server cluster, with one node having 10 regions
@@ -359,6 +371,8 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
       // validate slop checks without this override. We override just to ensure we will always
       // validate slop check here, and for small clusters as well.
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 1.0f);
+      conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+      conf.setLong("hbase.master.balancer.stochastic.maxSteps", 5000L);
       loadBalancer.onConfigurationChange(conf);
       for (int[] mockCluster : clusterStateMocksWithNoSlop) {
         assertTrue(hasEmptyBalancerPlans(mockCluster));
@@ -369,6 +383,8 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
     } finally {
       // reset config
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", minCost);
+      conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", runMaxSteps);
+      conf.setLong("hbase.master.balancer.stochastic.maxSteps", maxSteps);
       loadBalancer.onConfigurationChange(conf);
     }
   }
@@ -381,14 +397,21 @@ public class TestStochasticLoadBalancer extends StochasticBalancerTestBase {
       new int[] { 2, 8, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5,
         5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5 }, };
     float minCost = conf.getFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 0.025f);
+    boolean runMaxSteps = conf.getBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+    long maxSteps = conf.getLong("hbase.master.balancer.stochastic.maxSteps", 1000000);
+    boolean oldIsByTable = conf.getBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, false);
     try {
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", 1.0f);
+      conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", false);
+      conf.setLong("hbase.master.balancer.stochastic.maxSteps", 5000L);
       conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, true);
       loadBalancer.onConfigurationChange(conf);
       assertFalse(hasEmptyBalancerPlans(regionsPerServerPerTable));
     } finally {
       conf.setFloat("hbase.master.balancer.stochastic.minCostNeedBalance", minCost);
-      conf.unset(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE);
+      conf.setBoolean("hbase.master.balancer.stochastic.runMaxSteps", runMaxSteps);
+      conf.setLong("hbase.master.balancer.stochastic.maxSteps", maxSteps);
+      conf.setBoolean(HConstants.HBASE_MASTER_LOADBALANCE_BYTABLE, oldIsByTable);
       loadBalancer.onConfigurationChange(conf);
     }
   }