[HBASE-26989](https://issues.apache.org/jira/browse/HBASE-26989)

When running the tests locally, I see these runtime improvements:
`testNeedBalance`: from 120 seconds to 11 seconds
`testSloppyTablesLoadBalanceByTable` 27 seconds to <1 second
`testBalanceOfSloppyServers` 67 seconds to <1 second

I ran `testNeedBalance` 100 times locally, and the other two tests I ran 1000 times locally.

So total class `TestStochasticLoadBalancer` runtime reduces from 230 seconds to 31 seconds.

Additionally, we get more deterministic behavior, since tests are more likely to have consistent results with a max number of steps when compared to a max running time.